+++ api/responses/__init__.py
@@ -1,2 +1,3 @@
+from .tyre_impression_responses import tyre_impression_response
 from .response_wrapper import paginated_response, error_response
 from .tyre_model_responses import slim_tyre_model_response, tyre_model_response